Add files via upload
Homework for Unit 2. Lesson 2 - Inheritance
Make a class structure in python representing people in a school. Make a base class called Person, a class called Student, and another one called Teacher. Try to find as many methods and attributes as you can which belong to the different classes, and keep in mind which are common and which are not. For example name should be a Person attribute, while salary should only be available to Teacher.